% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_in_225.R
\name{people_in_225}
\alias{people_in_225}
\title{Show people in 225 office}
\usage{
people_in_225(people_number)
}
\arguments{
\item{people_number}{This is the number of people you want to see}
}
\value{
a character vector with length same as the people_number
}
\description{
Show people in 225 office
}
\examples{
people_in_225(1)
people_in_225(2)
people_in_225(3)
people_in_225(4)
}
